// src/models/common/IconType.ts
export enum IconType {
    faHome = "faHome",
    faBusinessTime = "faBusinessTime",
    faKey = "faKey",
    faGauge = "faGauge",
    faAccessibleIcon = "faAccessibleIcon",
    faUserTie = "faUserTie",
    faMoneyBillTrendUp = "faMoneyBillTrendUp",
    faFileCode = "faFileCode",
    faChessPawn = "faChessPawn",
    faChessBishop = "faChessBishop",
    faChessKnight = "faChessKnight",
    faChessRook = "faChessRook",
    faChessQueen = "faChessQueen",
    faChessKing = "faChessKing",
    faHandshake = "faHandshake",
    faSuitcase = "faSuitcase",
    faCube = "faCube",
    faChess = "faChess",
    faCompass = "faCompass",
    faAngleDown = "faAngleDown",
    faAddressCard = "faAddressCard",
    faLink = "faLink",
    faGithubAlt = "faGithubAlt",
    faPalette = "faPalette",
    faMoon = "faMoon",
    faSun = "faSun",
    faRotateLeft = "faRotateLeft",
    faChevronLeft = "faChevronLeft",
    faMobile = "faMobile",
    faEnvelope = "faEnvelope",
    faLinkedin = "faLinkedin",
    faXmark = "faXmark",
    faFaceSadCry = "faFaceSadCry",
    faHandshakeAngle = "faHandshakeAngle",
    faPeace = "faPeace",
    faStar = "faStar",
    faTrophy = "faTrophy",
    faSave = "faSave",
}